Reimplement `let` and `param` values as `static` methods + lambdas instead of inner classes.
This on its own is a minor improvement, but will unlock optimistic evaluation in a followup change.  In the mean time this will be a massive decrease in the number of classes generated and the overall size of those classes since we eliminate a ton of 'constant pool duplication'.

One tricky thing was reworking the naming algorithm for these 'inner methods'. For that we are using a hierarchical naming scheme where the name of the methods are a combination of the parent template and the local param/let variable.

PiperOrigin-RevId: 597331590